refactor(atlassian,workspace): split large files under 500-line threshold (#644)
